MVC tarafı ve RabbitMQ ayarları bitti.

+++ Frontends/FreeCourse.Web/Services/OrderService.cs
@@ -70,15 +70,46 @@ namespace FreeCourse.Web.Services
 
             var orderCreatedViewModel = await response.Content.ReadFromJsonAsync<Response<OrderCreatedViewModel>>();
             orderCreatedViewModel.Data.IsSuccessful = true;
-            _basketService.Delete();
+            await _basketService.Delete();
             return orderCreatedViewModel.Data;
 
 
         }
 
-        public Task SuspendOrder(CheckoutInfoInput checkoutInfoInput)
+        public async Task<OrderSuspendViewModel> SuspendOrder(CheckoutInfoInput checkoutInfoInput)
         {
-            throw new NotImplementedException();
+            var basket = await _basketService.Get();
+            var orderCreatedInput = new OrderCreateInput()
+            {
+                BuyerId = _sharedIdentityService.GetUserId,
+                Address = new AddressCreateInput() { District = checkoutInfoInput.District, Line = checkoutInfoInput.Line, Province = checkoutInfoInput.Province, Street = checkoutInfoInput.Street, ZipCode = checkoutInfoInput.ZipCode }
+
+            };
+            basket.BasketItems.ForEach(x =>
+            {
+                var orderItem = new OrderItemCreateInput() { ProductId = x.CourseId, Price = x.GetCurrentPrice, PictureUrl = "", ProductName = x.CourseName };
+                orderCreatedInput.OrderItems.Add(orderItem);
+            });
+            
+            var paymentInfoInput = new PaymentInfoInput()
+            {
+                CardName = checkoutInfoInput.CardName,
+                CardNumber = checkoutInfoInput.CardNumber,
+                Expiration = checkoutInfoInput.Expiration,
+                CVV = checkoutInfoInput.CVV,
+                TotalPrice = basket.TotalPrice,
+                Order = orderCreatedInput
+            };
+            var responsePayment = await _paymentService.ReceivePayment(paymentInfoInput);
+
+            if (!responsePayment)
+            {
+                return new OrderSuspendViewModel() { Error = "Ödeme alınamadı", IsSuccessful = false };
+            }
+
+            await _basketService.Delete();
+            return new OrderSuspendViewModel() {IsSuccessful = true};
+
         }
 
         public async Task<List<OrderViewModel>> GetOrder()
